add folds to rock faction blocks file
The diff is strange? Apparently this commit adds a carriage return to
the end of a couple comments? I don't know why that would happen, I
never typed in those characters NOR used github desktop, which seems to
really like to add carriage return characters to the end of lines.